Add independent Book Class, remove unnecessary eslint-disable lines
