+++ modules/BookRackClass.js
@@ -1,0 +1,11 @@
+export default class BooksRack {
+  constructor() {
+    this.Books = [];
+  }
+
+  UpdateRack() {
+    if (localStorage.Books) {
+      this.Books = JSON.parse(localStorage.Books);
+    }
+  }
+}
